Added solved conda environment and requirements.txt
Added a enviroment reference sheet to lock down the versions of
the environment and identified all immediately known dependencies.
Also added a readme to explain how to run the system using both
conda and python w/ pip.
Added a procedure in the readme to convert the jupyter notebook into
jupytext formatted py file, which will make it easier to simply run
the code without the jupyter overhead.