Update slstr_cloud.py
Inference updated, now it calculates accuracy. The accuracy is calculated by comparing the values of pixels of the mask produced by the mode and the ground truth. The ground truth is also a mask which contains Bayesian values. These values where converted to values of 0 (clear sky) and 1 (presence of a cloud).